CHG: refactor the folders; reduce the complexity of inference by using a general way for setting model up
