+++ sssd/core/imputers/SSSDSAImputer.py
@@ -3,7 +3,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 from einops import rearrange
 
-from sssd.imputers.S4Model import S4, LinearActivation
+from sssd.core.imputers.S4Model import S4, LinearActivation
 from sssd.utils.util import calc_diffusion_step_embedding
